requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.25.0
botocore==1.28.0
colorama==0.4.5
Django==4.1.2
django-filter==22.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
jmespath==1.0.1
Markdown==3.4.1
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
zipp==3.10.0
